[FE] modify: 세부 컴포넌트로 분리 수정
